% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone-package.R
\docType{package}
\name{zone-package}
\alias{zone}
\alias{zone-package}
\title{zone: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/zone}
  \item Report bugs at \url{https://github.com/DavisVaughan/zone/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item Howard Hinnant (Author of the included date library) [copyright holder]
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
