% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{zone_database_names}
\alias{zone_database_names}
\title{Time zone database names}
\usage{
zone_database_names()
}
\value{
A character vector of zone names.
}
\description{
\code{zone_database_names()} returns the time zone names found in the database.
}
\examples{
zone_database_names()
}
