% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{zone_database_version}
\alias{zone_database_version}
\title{Time zone database version}
\usage{
zone_database_version()
}
\value{
A single string of the database version.
}
\description{
\code{zone_database_version()} returns the version of the time zone database
currently in use.
}
\examples{
zone_database_version()
}
