% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones-package.R
\docType{package}
\name{zones-package}
\alias{zones}
\alias{zones-package}
\title{zones: 'date' C++ Library Access for Working With Dates and Times}
\description{
Provides C++ headers and callables for working with the
    'date' library. 'date' provides comprehensive support for working with
    dates and date-times, which this package exposes to make it easier for
    other R packages to utilize. Headers are provided for calendar
    specific calculations, along with a limited callable interface for
    time zone manipulations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/zones}
  \item Report bugs at \url{https://github.com/DavisVaughan/zones/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item Howard Hinnant (Author of the included date library) [copyright holder]
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
