% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{zones_initialize}
\alias{zones_initialize}
\title{Initialize zones for usage in other packages}
\usage{
zones_initialize()
}
\value{
\code{NULL}, invisibly.
}
\description{
\code{zones_initialize()} is intended to be called from a client package's
\code{.onLoad()} as \code{zones::zones_initialize()} to ensure that the zones package
has been loaded.

The function itself doesn't do anything. It is instead called for the
side-effect of loading the zones package. This does two things:
\itemize{
\item The zones \code{.onLoad()} hook is run, which sets the path to the time zone
database that exists in tzdb.
\item The callables in zones are registered, which allows them to be called from
other packages.
}
}
\details{
There are alternative ways to ensure that zones is loaded. A client package
can alternatively import a function from zones into its package with the
\verb{@importFrom} tag, or can call \code{requireNamespace("zones", quiet = TRUE)} from
its \code{.onLoad()}.
}
\examples{
zones_initialize()
}
